Refactored directory references
