+++ boosters/node-beginner/package.json
@@ -15,7 +15,8 @@
   "devDependencies": {},
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "nodemon ./src/app.js"
+    "start": "nodemon ./src/app.js",
+    "db:setup": "sequelize db:migrate"
   },
   "author": "",
   "license": "ISC"
