+++ boosters/node-beginner/src/api/routes/repos/index.js
@@ -1,9 +1,9 @@
-const express = require('express');
-const dbRepo = require('../../../models').repo;
+const express = require("express");
+const { Repo: dbRepo } = require("../../../../models");
 const router = express.Router();
 
 router
-  .route('/')
+  .route("/")
   .get(async (req, res) => {
     const repos = await dbRepo.findAll();
     res.status(200);
@@ -14,18 +14,18 @@ router
       await dbRepo.create({
         name: req.body.name,
         userId: req.body.userId,
-        isPrivate: req.body.isPrivate,
+        isPrivate: req.body.isPrivate
       });
       res.status(201);
-      res.end('Added repo');
+      res.end("Added repo");
     } catch (e) {
       res.status(400);
       console.error(e);
-      res.end('Invalid request');
+      res.end("Invalid request");
     }
   });
 
-router.get('/:repoId', async (req, res) => {
+router.get("/:repoId", async (req, res) => {
   const repo = await dbRepo.findByPk(req.params.repoId);
   res.send(repo);
 });
