+++ boosters/node-beginner/src/api/routes/users/index.js
@@ -1,10 +1,9 @@
-const express = require('express');
+const express = require("express");
 const router = express.Router();
-const dbUser = require('../../../models').user;
-const dbRepo = require('../../../models').repo;
+const { User: dbUser, Repo: dbRepo } = require("../../../../models");
 
 router
-  .route('/')
+  .route("/")
   .get(async (req, res) => {
     const users = await dbUser.findAll();
     res.send(users);
@@ -13,27 +12,27 @@ router
     try {
       await dbUser.create({
         firstName: req.body.name,
-        lastName: req.body.lastName,
+        lastName: req.body.lastName
       });
       res.status(201);
-      res.end('Added user');
+      res.end("Added user");
     } catch (e) {
       res.status(400);
       console.error(e);
-      res.end('Invalid request');
+      res.end("Invalid request");
     }
   });
 
-router.get('/:userId', async (req, res) => {
+router.get("/:userId", async (req, res) => {
   const user = await dbUser.findByPk(req.params.userId);
   res.send(user);
 });
 
-router.get('/:userId/repos', async (req, res) => {
+router.get("/:userId/repos", async (req, res) => {
   const repos = await dbRepo.findAll({
     where: {
-      userId: req.params.userId,
-    },
+      userId: req.params.userId
+    }
   });
   res.send(repos);
 });
