+++ boosters/node-beginner/src/app.js
@@ -1,18 +1,18 @@
-require('dotenv').config();
-const express = require('express');
-const app = express();
-const routes = require('./api/routes');
-const db = require('./models').db;
-const bodyParser = require('body-parser');
+require("dotenv").config();
+const express = require("express");
+const routes = require("./api/routes");
+const { db } = require("../models");
+const bodyParser = require("body-parser");
 
-db.sync({ force: true }).then(() => {
-  app.listen(process.env.port || 8080, () => {
-    console.log(`Server ready. Listening at http://localhost:8080`);
-  });
+const app = express();
+
+app.listen(process.env.port || 8080, () => {
+  console.log(`Server ready. Listening at http://localhost:8080`);
 });
+
 app.use(bodyParser.json());
-app.get('/', (req, res) => {
-  res.send('Hello World!');
+app.get("/", (req, res) => {
+  res.send("Hello World!");
 });
-app.use('/users', routes.userRoutes);
-app.use('/repos', routes.repoRoutes);
+app.use("/users", routes.userRoutes);
+app.use("/repos", routes.repoRoutes);
